fixup

+++ src/cuda-sim/ptx_sim.h
@@ -369,7 +369,7 @@ class ptx_thread_info {
   void set_ntid(dim3 tid) { m_ntid = tid; }
   void set_nctaid(dim3 cta_size) { m_nctaid = cta_size; }
 
-  unsigned get_builtin(int builtin_id, unsigned dim_mod);
+  unsigned long long get_builtin(int builtin_id, unsigned dim_mod);
 
   void set_done();
   bool is_done() { return m_thread_done; }
